Build:Update module structure
